requirements.txt
Django
Django-grappelli
Pillow
Django-imagekit
Django-filebrowser
Pytz
dj-static